Extract UserController#contact to Users::MessagesController#contact
